fixed many errors flagged by the 10.6 xcode compiler analysis tools
git-svn-id: svn://orca.physics.unc.edu/Orca@2412 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Main Dialogs/ORCARootService/ORCARootServiceController.m
@@ -259,12 +259,12 @@ SYNTHESIZE_SINGLETON_FOR_CLASS(ORCARootServiceController);
 	}
 }
 
-- (unsigned ) numberOfItemsInComboBox:(NSComboBox *)aComboBox
+- (NSInteger ) numberOfItemsInComboBox:(NSComboBox *)aComboBox
 {
 	return  [[self orcaRootService] connectionHistoryCount];
 }
 
-- (id)comboBox:(NSComboBox *)aComboBox objectValueForItemAtIndex:(unsigned)index
+- (id)comboBox:(NSComboBox *)aComboBox objectValueForItemAtIndex:(NSInteger)index
 {
 	return [[self orcaRootService] connectionHistoryItem:index];
 }
